jxl-render: Reorganize filters
