+++ crates/jxl-render/src/filter/impls.rs
@@ -1,5 +1,3 @@
-mod common;
-
 #[cfg(target_arch = "aarch64")]
 mod aarch64;
 mod generic;
