+++ crates/jxl-render/src/filter/impls/aarch64.rs
@@ -1,16 +1,15 @@
-use std::arch::aarch64::*;
 use std::arch::is_aarch64_feature_detected;
 
 use jxl_frame::{filter::EpfParams, FrameHeader};
 use jxl_grid::SimpleGrid;
 use jxl_threadpool::JxlThreadPool;
 
+use crate::filter::epf::run_epf_rows;
 use crate::Region;
 use crate::Result;
 
-use super::generic::epf_common;
-
 mod epf;
+mod gabor;
 
 pub fn epf<const STEP: usize>(
     input: &[SimpleGrid<f32>; 3],
@@ -24,7 +23,7 @@ pub fn epf<const STEP: usize>(
     if is_aarch64_feature_detected!("neon") {
         // SAFETY: Features are checked above.
         unsafe {
-            return epf_common(
+            return run_epf_rows(
                 input,
                 output,
                 frame_header,
@@ -40,7 +39,7 @@ pub fn epf<const STEP: usize>(
 
     // SAFETY: row handler is safe.
     unsafe {
-        epf_common(
+        run_epf_rows(
             input,
             output,
             frame_header,
@@ -59,7 +58,7 @@ pub fn apply_gabor_like(fb: [&mut SimpleGrid<f32>; 3], weights_xyb: [[f32; 2]; 3
         // SAFETY: Features are checked above.
         unsafe {
             for (fb, [weight1, weight2]) in fb.into_iter().zip(weights_xyb) {
-                run_gabor_inner_neon(fb, weight1, weight2)?
+                gabor::run_gabor_inner_neon(fb, weight1, weight2)?
             }
         }
         return Ok(());
@@ -67,131 +66,6 @@ pub fn apply_gabor_like(fb: [&mut SimpleGrid<f32>; 3], weights_xyb: [[f32; 2]; 3
 
     for (fb, [weight1, weight2]) in fb.into_iter().zip(weights_xyb) {
         super::generic::run_gabor_inner(fb, weight1, weight2)?;
-    }
-    Ok(())
-}
-
-#[target_feature(enable = "neon")]
-unsafe fn run_gabor_inner_neon(fb: &mut SimpleGrid<f32>, weight1: f32, weight2: f32) -> Result<()> {
-    let global_weight = (1.0 + weight1 * 4.0 + weight2 * 4.0).recip();
-
-    let width = fb.width();
-    let height = fb.height();
-    if width * height <= 1 {
-        return Ok(());
-    }
-
-    let tracker = fb.tracker();
-    let io = fb.buf_mut();
-
-    let _handle = tracker
-        .as_ref()
-        .map(|x| x.alloc::<f32>(width))
-        .transpose()?;
-    let mut prev_row = io[..width].to_vec();
-
-    let input_t = prev_row.as_mut_ptr();
-    let mut input_c = io.as_mut_ptr();
-    let mut input_b = input_c.add(width);
-
-    for _ in 0..height - 1 {
-        let mut tl = vld1_dup_f32(input_t);
-        let mut cl = vld1_dup_f32(input_c);
-        let mut bl = vld1_dup_f32(input_b);
-        *input_t = *input_c;
-
-        for vx in 0..(width - 1) / 2 {
-            let tr = vld1_f32(input_t.add(1 + vx * 2));
-            let cr = vld1_f32(input_c.add(1 + vx * 2));
-            let br = vld1_f32(input_b.add(1 + vx * 2));
-            let t = vext_f32::<1>(tl, tr);
-            let c = vext_f32::<1>(cl, cr);
-            let b = vext_f32::<1>(bl, br);
-
-            let sum_side = vadd_f32(vadd_f32(vadd_f32(t, cl), cr), b);
-            let sum_diag = vadd_f32(vadd_f32(vadd_f32(tl, tr), bl), br);
-            let unweighted_sum = vfma_n_f32(vfma_n_f32(c, sum_side, weight1), sum_diag, weight2);
-            let sum = vmul_n_f32(unweighted_sum, global_weight);
-
-            vst1_f32(input_t.add(1 + vx * 2), cr);
-            vst1_f32(input_c.add(vx * 2), sum);
-            tl = tr;
-            cl = cr;
-            bl = br;
-        }
-
-        let tr;
-        let cr;
-        let br;
-        if width % 2 == 0 {
-            tr = vld1_dup_f32(input_t.add(width - 1));
-            cr = vld1_dup_f32(input_c.add(width - 1));
-            br = vld1_dup_f32(input_b.add(width - 1));
-        } else {
-            tr = vdup_lane_f32::<1>(tl);
-            cr = vdup_lane_f32::<1>(cl);
-            br = vdup_lane_f32::<1>(bl);
-        };
-        let t = vext_f32::<1>(tl, tr);
-        let c = vext_f32::<1>(cl, cr);
-        let b = vext_f32::<1>(bl, br);
-
-        let sum_side = vadd_f32(vadd_f32(vadd_f32(t, cl), cr), b);
-        let sum_diag = vadd_f32(vadd_f32(vadd_f32(tl, tr), bl), br);
-        let unweighted_sum = vfma_n_f32(vfma_n_f32(c, sum_side, weight1), sum_diag, weight2);
-        let sum = vmul_n_f32(unweighted_sum, global_weight);
-
-        if width % 2 == 0 {
-            *input_t.add(width - 1) = vget_lane_f32::<0>(cr);
-            vst1_f32(input_c.add(width - 2), sum);
-        } else {
-            *input_c.add(width - 1) = vget_lane_f32::<0>(sum);
-        }
-
-        input_c = input_c.add(width);
-        input_b = input_b.add(width);
-    }
-
-    let mut tl = vld1_dup_f32(input_t);
-    let mut cl = vld1_dup_f32(input_c);
-
-    for vx in 0..(width - 1) / 2 {
-        let tr = vld1_f32(input_t.add(1 + vx * 2));
-        let cr = vld1_f32(input_c.add(1 + vx * 2));
-        let t = vext_f32::<1>(tl, tr);
-        let c = vext_f32::<1>(cl, cr);
-
-        let sum_side = vadd_f32(vadd_f32(vadd_f32(t, cl), cr), c);
-        let sum_diag = vadd_f32(vadd_f32(vadd_f32(tl, tr), cl), cr);
-        let unweighted_sum = vfma_n_f32(vfma_n_f32(c, sum_side, weight1), sum_diag, weight2);
-        let sum = vmul_n_f32(unweighted_sum, global_weight);
-
-        vst1_f32(input_c.add(vx * 2), sum);
-        tl = tr;
-        cl = cr;
-    }
-
-    let tr;
-    let cr;
-    if width % 2 == 0 {
-        tr = vld1_dup_f32(input_t.add(width - 1));
-        cr = vld1_dup_f32(input_c.add(width - 1));
-    } else {
-        tr = vdup_lane_f32::<1>(tl);
-        cr = vdup_lane_f32::<1>(cl);
-    };
-    let t = vext_f32::<1>(tl, tr);
-    let c = vext_f32::<1>(cl, cr);
-
-    let sum_side = vadd_f32(vadd_f32(vadd_f32(t, cl), cr), c);
-    let sum_diag = vadd_f32(vadd_f32(vadd_f32(tl, tr), cl), cr);
-    let unweighted_sum = vfma_n_f32(vfma_n_f32(c, sum_side, weight1), sum_diag, weight2);
-    let sum = vmul_n_f32(unweighted_sum, global_weight);
-
-    if width % 2 == 0 {
-        vst1_f32(input_c.add(width - 2), sum);
-    } else {
-        *input_c.add(width - 1) = vget_lane_f32::<0>(sum);
     }
     Ok(())
 }
